web/backend/django/django-crud/requirements.txt
astroid==2.2.5
backcall==0.1.0
beautifulsoup4==4.8.0
colorama==0.4.1
decorator==4.4.0
Django==2.2.5
django-appconf==1.0.3
django-bootstrap4==1.0.1
django-extensions==2.2.1
django-imagekit==4.0.2
Faker==2.0.1
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
parso==0.5.1
pickleshare==0.7.5
pilkit==2.0
Pillow==6.1.0
prompt-toolkit==2.0.9
Pygments==2.4.2
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
